backend/requirements.txt
flask
google-cloud-firestore
google-cloud-storage
python-dotenv
werkzeug